Found two major performance issues:
* Step function lookup is *still* very slow, which probably explains why the interpreter was spending about 25% of its time in the `run` function.  Fixed this by looking up the step function in the `State` constructor, so it only needs to be done once per state object instead of once per step.
* For some reason, one of the serialisation test in the internal test harness somehow poisons node.js's state in a way that causes subsequent benchmarks (not using the same `Interpreter` instance) to run more slowly.  This explains the observed discrepancy between the test harness benchmark and the test db benchmark (which runs in a separate instance of node.js) that has existed for some time.

The poisonous test is useful for maintaining correctness, so I have left it in for now.  The fact that it causes serious performance penalties merits further investigation of the mechanism involved, rather than just tweaking the test until it doesn't; to that end I have filed bug #193.

Best observed performance on Fibonacci benchmark:

Condition | `benchFibbonacci10k` (test harness) | `test_fibonacci10k` (test database)
----------|---------------------------------------|------------------------------------
Before this PR | 2889ms | 2520ms
After this PR | 2307ms | 1909ms
After removing poisonous test | 1967ms | 1985ms